H20EZE 02.10 is partial better to 02.13, keep both
If we look at Release_Note.txt inside the archive, we could see all the differences between 02.10 and 02.13.
Although 02.13 adds Tiano decompress support and a few extra features / tweaks, version 02.10 is partially better:
1) 02.10 has "SetupEditor" in File menu, while it has been removed since 02.11 so 02.13 doesn't have it
2) 02.10 shows descriptions for more BIOS modules in "Components --> Modules --> Module Export" menu
e.g. I open one particular BIOS in 02.10 and see that it has a lot of text descriptions:
...
0325B5A1-0937-4A4F-B8AF-EC3F80EE6B35 (SataController)
...
07A9330A-F347-11D4-9A49-0090273FC14D (LegacyMetronome)
...
0997E770-756E-4139-BF3F-7FC2A308177E (SmmS3)
...
and so on. Meanwhile, from the examples above, 02.13 has text description only for 0997E770-756E-4139-BF3F-7FC2A308177E - it is SmmS3
Descriptions are very valuable for understanding the BIOS functions, and so far 02.13 loses to 02.10 in this relation
I believe that eac